chore: better organize finance domain
moved dashboard and monitor views to finance domain
