+++ src/scripts/finance/views/dashboard/dashboard.js
