+++ src/scripts/finance/views/monitor/monitor.js
