+++ src/scripts/finance/views/monitor/monitor.test.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { shallow } from 'enzyme';
 import { WViewport } from '@scripts/base/components/viewport/viewport';
 import { WFinanceMonitorCard } from '@scripts/finance/components/finance-monitor-card/finance-monitor-card';
-import { Monitor } from '@scripts/monitor/monitor';
+import { Monitor } from './monitor';
 
 describe('Monitor', () => {
   function mount(){
